Move more code from unit.cpp
Delete unit_damage.h
